Refactor v2 parser condition tests
